Move PyPI and requirements files tests into different folder
